feat(types): add tests for types
The tests were based on fastify-jwt and other projects that have
type definition and types test. The test will cover types on each of the
modules: 'assert', 'errorHandler', 'forwarded', 'httpErrors', 'is',
'proxyaddr', 'to' and 'vary' (coverage was based on current JS tests).

As only the types will be checked, `tsc` is enough for that.